Move stateLens utils to utils/react
