+++ src/themes/greyVest/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import F from 'futil-js'
 import { observer } from 'mobx-react'
 import { defaultProps } from 'recompose'
-import { useLens } from '../../utils/futil'
+import { useLens } from '../../utils/react'
 import { withNode, contexturify } from '../../utils/hoc'
 import {
   Flex,
