+++ stories/imdb/stories/checklist.js
@@ -5,7 +5,7 @@ import { observable } from 'mobx'
 import { fromPromise } from 'mobx-utils'
 import { observer } from 'mobx-react'
 import Contexture, { updateSchemas } from '../utils/contexture'
-import { useLens } from '../../../src/utils/futil'
+import { useLens } from '../../../src/utils/react'
 import { FilterList, Flex, Awaiter, SpacedList, Grid } from '../../../src'
 import * as Theme from '../../../src/themes/greyVest'
 let { Adder, Button, Pager, ExampleTypes, ButtonRadio } = Theme
